rej_uniform_eta4 scalar-tail: fix R10_NIBBLE_VAL tower shape to match VSTEPS
The X86_VSTEPS-emitted low-nibble zx-tower is byte->int32->int64->int32->int64->int32
(alternating widths from movzbl semantics), not the byte->int16->int32 shape first
guessed. Corrected so R10_NIBBLE_VAL REWRITEs the in-context R10 read to
word(val(EL p inlist) MOD 16). Lesson: prove these as standalone lemmas (clean
prove(), no ambient asms) then REWRITE in-context — an in-context SIMP with
VAL_WORD_ZX_GEN loops.

+++ proofs/hol_light/x86_64/proofs/.scalar_tail_build.ml
@@ -102,7 +102,7 @@ let RESOLVE_RIP_FAST =
    word(val b MOD 16):int64. The zx-tower shape (byte->int16->int32 widenings)
    is exactly what VSTEPS emits for the load + zero-extends. *)
 let R10_NIBBLE_VAL = prove
- (`!b:byte. word_zx(word_and (word_zx (word_zx (word_zx (word_zx (word_zx b:int16):int32):int32):int32):int32) (word 15:int32)):int64 = word(val b MOD 16)`,
+ (`!b:byte. word_zx(word_and (word_zx (word_zx (word_zx (word_zx (word_zx b:int32):int64):int32):int64):int32) (word 15:int32)):int64 = word(val b MOD 16)`,
   GEN_TAC THEN REWRITE_TAC[GSYM VAL_EQ] THEN
   SUBGOAL_THEN `(word 15:int32) = word(2 EXP 4 - 1)` SUBST1_TAC THENL
    [CONV_TAC NUM_REDUCE_CONV; ALL_TAC] THEN
